Mehrere Einträge gleichzeigit löschen (#355)
* Mehrere Einträge gleichzeigit löschen bei: Mitglieder, Nicht-Mitglieder,
Kursteilnehmer, Emails, Emailempfänger, Buchungsarten
"Mail an Spender" auch bei Mehrfachasuwahl
MailVorlageDialog Doppelklick -> übernehmen statt bearbeiten

* Import entfernt

+++ src/de/jost_net/JVerein/gui/action/BuchungsartDeleteAction.java
@@ -33,21 +33,29 @@ public class BuchungsartDeleteAction implements Action
   @Override
   public void handleAction(Object context) throws ApplicationException
   {
-    if (context == null || !(context instanceof Buchungsart))
+    Buchungsart[] buchungsarten = null;
+    if (context == null)
     {
       throw new ApplicationException("Keine Buchungsart ausgewählt");
     }
+    else if(context instanceof Buchungsart)
+    {
+      buchungsarten = new Buchungsart[] {(Buchungsart)context};
+    }
+    else if(context instanceof Buchungsart[])
+    {
+      buchungsarten = (Buchungsart[])context;
+    }
+    else
+    {
+      return;
+    }
     try
     {
-      Buchungsart b = (Buchungsart) context;
-      if (b.isNewObject())
-      {
-        return;
-      }
-      
+      String mehrzahl = buchungsarten.length > 1? "en":"";
       YesNoDialog d = new YesNoDialog(YesNoDialog.POSITION_CENTER);
-      d.setTitle("Buchungsart löschen");
-      d.setText("Wollen Sie diese Buchungsart wirklich löschen?");
+      d.setTitle("Buchungsart" + mehrzahl +" löschen");
+      d.setText("Wollen Sie diese Buchungsart" + mehrzahl + " wirklich löschen?");
       try
       {
         Boolean choice = (Boolean) d.open();
@@ -60,8 +68,13 @@ public class BuchungsartDeleteAction implements Action
         return;
       }
 
-      b.delete();
-      GUI.getStatusBar().setSuccessText("Buchungsart gelöscht.");
+      for(Buchungsart b:buchungsarten)
+      {
+        if(b.isNewObject())
+          continue;
+        b.delete();
+      }
+      GUI.getStatusBar().setSuccessText("Buchungsart" + mehrzahl + " gelöscht.");
     }
     catch (RemoteException e)
     {
